requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langsmith==0.1.69
multidict==6.0.5
numpy==1.26.4
openai==1.31.0
orjson==3.10.3
packaging==23.2
PuLP==2.8.0
pydantic==2.7.3
pydantic_core==2.18.4
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
yarl==1.9.4
